Rename `rustc_expand::build` to `builder`
