+++ compiler/rustc_expand/src/lib.rs
@@ -18,7 +18,7 @@
 
 extern crate proc_macro as pm;
 
-mod build;
+mod builder;
 mod errors;
 // FIXME(Nilstrieb) Translate macro_rules diagnostics
 #[allow(rustc::untranslatable_diagnostic)]
